Revert "health: move cilium health logic from cilium-health/launch to pkg/health"
This reverts commit a6bc06e01d05005a58862f410ec63b7ef898be36.
