+++ cilium-health/launch/endpoint.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package health
+package launch
 
 import (
 	"context"
